# front-land/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/roommates/get-all-roommates": {
      "get": {
        "tags": [
          "Roommates"
        ],
        "operationId": "GetAllRoommates",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minBudget",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "maxBudget",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "smokingAllowed",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "petFriendly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lifestyle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profession",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availableFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "stayDuration",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "apartmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RoommateGender"
            }
          },
          {
            "name": "workSchedule",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/WorkSchedule"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/roommates/get-roommate": {
      "get": {
        "tags": [
          "Roommates"
        ],
        "operationId": "GetRoommate",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roommates/get-roommate-by-user-id": {
      "get": {
        "tags": [
          "Roommates"
        ],
        "operationId": "GetRoommateByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roommates/create-roommate": {
      "post": {
        "tags": [
          "Roommates"
        ],
        "operationId": "CreateRoommate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roommates/update-roommate/{id}": {
      "put": {
        "tags": [
          "Roommates"
        ],
        "operationId": "UpdateRoommate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoommateDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roommates/delete-roommate/{id}": {
      "delete": {
        "tags": [
          "Roommates"
        ],
        "operationId": "DeleteRoommate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RoommateDto": {
        "type": "object",
        "properties": {
          "roommateId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "profilePicture": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "hobbies": {
            "type": "string",
            "nullable": true
          },
          "profession": {
            "type": "string",
            "nullable": true
          },
          "smokingAllowed": {
            "type": "boolean",
            "nullable": true
          },
          "petFriendly": {
            "type": "boolean",
            "nullable": true
          },
          "lifestyle": {
            "type": "string",
            "nullable": true
          },
          "cleanliness": {
            "type": "string",
            "nullable": true
          },
          "guestsAllowed": {
            "type": "boolean",
            "nullable": true
          },
          "budgetMin": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "budgetMax": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "budgetIncludes": {
            "type": "string",
            "nullable": true
          },
          "availableFrom": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "availableUntil": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "minimumStayMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maximumStayMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lookingForRoomType": {
            "type": "string",
            "nullable": true
          },
          "lookingForApartmentType": {
            "type": "string",
            "nullable": true
          },
          "preferredLocation": {
            "type": "string",
            "nullable": true
          },
          "lookingForApartmentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "gender": {
            "$ref": "#/components/schemas/RoommateGender"
          },
          "languages": {
            "type": "string",
            "nullable": true
          },
          "workSchedule": {
            "$ref": "#/components/schemas/WorkSchedule"
          },
          "musicFriendly": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoommateGender": {
        "enum": [
          "PreferNotToSay",
          "Male",
          "Female",
          "NonBinary",
          "Other"
        ],
        "type": "string"
      },
      "RoommateInputDto": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "nullable": true
          },
          "hobbies": {
            "type": "string",
            "nullable": true
          },
          "profession": {
            "type": "string",
            "nullable": true
          },
          "smokingAllowed": {
            "type": "boolean",
            "nullable": true
          },
          "petFriendly": {
            "type": "boolean",
            "nullable": true
          },
          "lifestyle": {
            "type": "string",
            "nullable": true
          },
          "cleanliness": {
            "type": "string",
            "nullable": true
          },
          "guestsAllowed": {
            "type": "boolean",
            "nullable": true
          },
          "budgetMin": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "budgetMax": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "budgetIncludes": {
            "type": "string",
            "nullable": true
          },
          "availableFrom": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "availableUntil": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "minimumStayMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maximumStayMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lookingForRoomType": {
            "type": "string",
            "nullable": true
          },
          "lookingForApartmentType": {
            "type": "string",
            "nullable": true
          },
          "preferredLocation": {
            "type": "string",
            "nullable": true
          },
          "lookingForApartmentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/RoommateGender"
          },
          "languages": {
            "type": "string",
            "nullable": true
          },
          "workSchedule": {
            "$ref": "#/components/schemas/WorkSchedule"
          },
          "musicFriendly": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WorkSchedule": {
        "enum": [
          "Flexible",
          "Morning",
          "Evening",
          "Night"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Please enter JWT with Bearer into field (e.g. 'Bearer {token}')",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}